feat(ast): implement CanonicalTypeTransformer and TypeTransformerBase for type transformations
